requirements.txt
requests>=2.25.1,<3.0.0
python-dotenv>=0.21.0,<1.0.0
python-socketio>=5.5.0,<6.0.0
